# docs/swagger.yaml
basePath: /
definitions:
  auth.AuthRequest:
    properties:
      password:
        description: |-
          Password of the user.
          example: strongpassword123
        type: string
      username:
        description: |-
          Username of the user.
          example: alice
        type: string
    type: object
  auth.AuthResponse:
    properties:
      token:
        description: |-
          JWT access token.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  lobbies.LobbyCreationResponse:
    properties:
      error:
        type: string
      lobby_id:
        type: string
    type: object
  lobbies.LobbyListItem:
    properties:
      id:
        description: |-
          Unique lobby identifier.
          example: lobby-1234
        type: string
      name:
        description: |-
          Display name of the lobby.
          example: Casual Game
        type: string
      playerCount:
        description: |-
          Number of players currently in the lobby.
          example: 3
        type: integer
    type: object
  lobbies.LobbyListResponse:
    properties:
      hasMore:
        description: |-
          Indicates whether more items are available.
          example: true
        type: boolean
      items:
        description: List of lobby items.
        items:
          $ref: '#/definitions/lobbies.LobbyListItem'
        type: array
      limit:
        description: |-
          Maximum number of items returned.
          example: 20
        type: integer
      offset:
        description: |-
          Offset used for pagination.
          example: 0
        type: integer
    type: object
host: localhost:1337
info:
  contact: {}
  description: This is the API for Munchin game backend
  title: Munchin API
  version: "0.1"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user using username and password and return a JWT
        token.
      parameters:
      - description: Authentication credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Invalid credentials or malformed request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Authenticate a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username and password and return a JWT
        token.
      parameters:
      - description: Registration credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Registration successful
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Invalid request or user already exists
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Register a new user
      tags:
      - auth
  /healthz:
    get:
      description: Works as a probe like healhtz check
      responses:
        "204":
          description: No Content
      summary: Check health of the service
      tags:
      - health
  /lobby:
    get:
      description: Retrieve a paginated list of lobbies for the lobby selection scene
        (Game Endpoint).
      parameters:
      - default: 20
        description: Maximum number of items to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of lobbies
          schema:
            $ref: '#/definitions/lobbies.LobbyListResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List lobbies
      tags:
      - lobby
    post:
      description: Create a new game lobby and initialize its game room for the authenticated
        player.
      produces:
      - application/json
      responses:
        "201":
          description: Lobby successfully created
          schema:
            $ref: '#/definitions/lobbies.LobbyCreationResponse'
        "400":
          description: Invalid request or unknown lobby
          schema:
            $ref: '#/definitions/lobbies.LobbyCreationResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new lobby
      tags:
      - lobby
  /lobby/{id}/join:
    post:
      description: Join an existing game lobby as the authenticated player.
      parameters:
      - description: Lobby ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Successfully joined the game
          schema:
            type: string
        "400":
          description: Missing lobby ID or invalid request
          schema:
            type: string
        "404":
          description: Lobby not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Join a game
      tags:
      - game
  /lobby/{id}/start:
    post:
      description: Start the game associated with a lobby.
      parameters:
      - description: Lobby ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Missing lobby ID or invalid lobby
          schema:
            $ref: '#/definitions/lobbies.LobbyCreationResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Start a game
      tags:
      - game
  /lobby/model:
    get:
      description: Retrieve all lobbies without pagination.
      produces:
      - application/json
      responses:
        "200":
          description: List of lobbies
          schema:
            items:
              $ref: '#/definitions/lobbies.LobbyListItem'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all lobbies
      tags:
      - lobby
swagger: "2.0"